slaunch_skinit: use relocator, relinquish TPM locality, send INIT to APs
Relocator must be used as it is the proper way of relocating code to
its final addresses. For now go with relocator32_boot pointing to skinit
stub in C function and pray that it won't get overwritten. This will be
fixed in the following commit.

TPM must have all localities relinquished and all APs must be in the
wait-for-SIPI state. These are requirements for skinit.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>