i386/msr: Fixes for MSR reading, writing and detection
MSR fixes. It does not make sense to have separate headers for separate static functions.
Additionally, we have to add some constants with MSR addresses in subsequent patches. So,
make one common place to store them.

More MSR fixes. Currently rdmsr and wrmsr commands have own MSR support detection code. This
code is the same. So, it is duplicated. Additionally, this code cannot be reused by others.
Hence, extract this code to a function and make it public. By the way, improve a code a bit.

Additionally, use GRUB_ERR_BAD_DEVICE instead of GRUB_ERR_BUG to signal
an error because errors encountered by this new routine are not bugs.

Signed-off-by: Daniel Kiper <daniel.kiper@oracle.com>