grub-core/loader/i386/txt/txt.c: update capabilities comparison logic
In the latest TXT Software Development Guide as of now (April 2023,
Revision 017.4) bits 4 and 5 (used to be "no legacy PCR usage" and
"auth PCR usage" respectively) of capabilities field bit are reserved.
Bit 4 is ignored, but it is returned as 0 by SINIT[CAPABILITIES],
while bit 5 is forced to 1 for compatibility reasons. This is related
to support for TPM 1.2 devices, which always had this bit set.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>