loader/multiboot_mbi2.c: disable UEFI BS when booting with slaunch
MLE (kernel or Xen) can't use pre-DLE services, as those are not
measured as part of DRTM. To make sure this is true, disable boot
services before MBI is created, which will ensure that all required
data is obtained and saved in MBI while it is available.

BS are disabled before call to grub_multiboot2_get_mbi_size(), so
the amount of allocated memory will be sufficient.

Signed-off-by: Krystian Hebel <krystian.hebel@3mdeb.com>